+++ PSCouchDB/PSCouchDB.psm1
@@ -759,7 +759,7 @@ class PSCouchDBDesignDoc : PSCouchDBDocument {
 
     # Constructor
     PSCouchDBDesignDoc () {
-        $this._id = "_design/$((New-CouchDBUuids -Count 1).uuids[0])"
+        $this._id = "_design/$((New-Guid).Guid.Replace('-', $null))"
         $this.views = New-Object Collections.Generic.List[PSCouchDBView]
         $this.doc['_id'] = $this._id
         $this.doc.Add('views', @{})
@@ -1093,6 +1093,7 @@ class PSCouchDBReplication {
     hidden [string] $selector
     hidden [string] $since_seq
     hidden [bool] $use_checkpoints
+    hidden [bool] $winning_revs_only
     hidden [hashtable] $replicator = @{}
 
     # Constuctor
@@ -1227,6 +1228,11 @@ class PSCouchDBReplication {
     UseCheckpoints () { 
         $this.use_checkpoints = $true
         $this.replicator.Add('use_checkpoints', $this.use_checkpoints)
+    }
+
+    SetWinningRevisionOnly ([bool]$value) {
+        $this.winning_revs_only = $value
+        $this.replicator.Add('winning_revs_only', $this.winning_revs_only)
     }
 
     [hashtable] GetDocument () {
@@ -1714,17 +1720,22 @@ function Send-CouchDBRequest {
         [pscredential] $ProxyCredential
     )
     # Create PSCouchDBRequest object
+    Write-Debug -Message "Create PSCouchDBRequest object"
     $req = New-Object PSCouchDBRequest
     $parameters = @()
     # Set cache
+    Write-Debug -Message "Set cache"
     if ($Global:CouchDBCachePreference) {
         $req.EnableCache()
     }
+    # Set server
+    Write-Debug -Message "Set server"
     if ($Server) {
         Write-Verbose -Message "Set server to $Server"
         $req.SetServer($Server)
     }
     # Set proxy server
+    Write-Debug -Message "Set proxy server"
     if ($ProxyServer) {
         if ($ProxyCredential -is [pscredential]) {
             Write-Verbose -Message "Set proxy server $ProxyServer with credential"
@@ -1735,17 +1746,20 @@ function Send-CouchDBRequest {
         }
     }
     # Set protocol
+    Write-Debug -Message "Set protocol"
     if ($Ssl.IsPresent) {
         # Set default port
         Write-Verbose -Message "Enable SSL on port 6984"
         $req.SetSsl()
     }
     # Set port
+    Write-Debug -Message "Set port"
     if ($Port) {
         Write-Verbose -Message "Set port on $Port"
         $req.SetPort($Port)
     }
     # Set method
+    Write-Debug -Message "Set method"
     switch ($Method) {
         "HEAD" { $req.SetMethod("HEAD") }
         "GET" { $req.SetMethod("GET") }
@@ -1756,6 +1770,7 @@ function Send-CouchDBRequest {
     }
     Write-Verbose -Message "Set method to $($req.method)"
     # Set database
+    Write-Debug -Message "Set database"
     if ($Database) {
         if ($Database -match "\?") {
             $arr = $Database -split "\?"
@@ -1769,6 +1784,7 @@ function Send-CouchDBRequest {
         $req.SetDatabase($Database)
     }
     # Set document
+    Write-Debug -Message "Set document"
     if ($Document) {
         if ($Document -match "\?") {
             $arr = $Document -split "\?"
@@ -1782,6 +1798,7 @@ function Send-CouchDBRequest {
         $req.SetDocument($Document)
     }
     # Set attachment
+    Write-Debug -Message "Set attachment"
     if ($Attachment) {
         if ('GET','HEAD' -contains $Method) {
             Write-Verbose -Message "Set attachment to $Attachment"
@@ -1796,11 +1813,13 @@ function Send-CouchDBRequest {
         }
     }
     # Set revision
+    Write-Debug -Message "Set revision"
     if ($Revision) {
         Write-Verbose -Message "Set revision to $Revision"
         $parameters += "rev=$Revision"
     }
     # Check other params
+    Write-Debug -Message "Set other params"
     if ($Params) {
         $parameters += $Params
     }
@@ -1809,6 +1828,7 @@ function Send-CouchDBRequest {
         $req.SetParameter($parameters)
     }
     # Check authorization
+    Write-Debug -Message "Set authorization"
     if ($Authorization -or $Global:CouchDBCredential) {
         if ($Global:CouchDBSaveCredentialPreference) {
             if (-not($Global:CouchDBCredential)) {
@@ -1828,11 +1848,13 @@ function Send-CouchDBRequest {
         }
     }
     # Check json data
+    Write-Debug -Message "Set json data"
     if ($Data) {
         Write-Verbose -Message "Set json data to: $Data"
         $req.SetData($Data)
     }
     # Get header or request
+    Write-Debug -Message "Set header or request"
     Write-Verbose -Message "Request to CouchDB server: $req"
     if ($Method -eq "HEAD") {
         $req.GetHeader()